requirements.txt
# FastAPI and ASGI
fastapi==0.119.1
uvicorn[standard]==0.38.0
python-multipart==0.0.20

# Redis
redis==6.4.0

# Gemini AI
google-generativeai==0.8.5

# Configuration
pydantic==2.12.3
pydantic-settings==2.11.0

# Development (optional)
pytest==8.4.2
pytest-asyncio==1.2.0
